+++ libs/core/src/lib/api/sdk.service.ts
@@ -5,6 +5,7 @@ import {
   IKnowledgeBoxItem,
   KnowledgeBox,
   LearningConfiguration,
+  LearningConfigurationSchemas,
   LearningConfigurationSet,
   Nuclia,
   WritableKnowledgeBox,
@@ -30,19 +31,13 @@ import { StateService } from '../state.service';
 import { FeatureFlagService } from '../analytics/feature-flag.service';
 import { take } from 'rxjs/operators';
 
-const IMMUTABLE_OPTIONS = ['semantic_model', 'visual_labeling'];
-
-export interface VisibleLearningConfiguration {
-  id: string;
-  data: LearningConfigurationWithFields;
-}
-
-export interface LearningConfigurationWithFields extends LearningConfiguration {
-  options: {
-    value: string;
-    name: string;
-    fields: { value: string; name: string }[];
-  }[];
+export interface LearningConfigurationUserKeys {
+  [key: string]: {
+    [key: string]: {
+      title: string;
+      required: boolean;
+    };
+  };
 }
 
 @Injectable({ providedIn: 'root' })
@@ -222,9 +217,11 @@ export class SDKService {
       .subscribe();
   }
 
-  getVisibleLearningConfiguration(
-    onCreation = true,
-  ): Observable<{ display: VisibleLearningConfiguration[]; full: LearningConfigurationSet }> {
+  getVisibleLearningConfiguration(onCreation = true): Observable<{
+    display: LearningConfigurationSet;
+    full: LearningConfigurationSet;
+    keys: LearningConfigurationUserKeys;
+  }> {
     return forkJoin([
       this.featureFlagService.isFeatureEnabled('kb-anonymization').pipe(take(1)),
       this.featureFlagService.isFeatureEnabled('answers').pipe(take(1)),
@@ -234,21 +231,37 @@ export class SDKService {
     ]).pipe(
       map(([hasAnonymization, hasAnswers, hasPdfAnnotation, conf, account]) => {
         const full = Object.entries(conf)
+          .filter(([, value]) => 'options' in value)
+          .map((entry) => entry as [string, LearningConfiguration])
           .map(([id, data]) => ({ id, data }))
           // some options cannot be changed after kb creation
-          .filter((entry) => onCreation || !IMMUTABLE_OPTIONS.includes(entry.id));
+          .filter((entry) => (onCreation && entry.data.create) || (!onCreation && entry.data.update));
 
         return {
           // At display, hide configurations with only one option or under feature flagging
           display: full.filter(
             (entry) =>
-              entry.data.options &&
               entry.data.options.length > 1 &&
               (entry.id !== 'anonymization_model' || hasAnonymization) &&
               (entry.id !== 'visual_labeling' || hasPdfAnnotation) &&
               (entry.id !== 'generative_model' || hasAnswers),
           ),
           full,
+          keys:
+            account.type === 'stash-enterprise'
+              ? Object.entries((conf['user_keys'] as LearningConfigurationSchemas)?.schemas || {}).reduce(
+                  (acc, [schemaId, schema]) => {
+                    acc[schemaId] = Object.entries(schema.properties)
+                      .filter(([, property]) => property.type === 'string')
+                      .reduce((acc, [propertyId, property]) => {
+                        acc[propertyId] = { title: property.title, required: schema.required.includes(propertyId) };
+                        return acc;
+                      }, {} as { [key: string]: { title: string; required: boolean } });
+                    return acc;
+                  },
+                  {} as LearningConfigurationUserKeys,
+                )
+              : {},
         };
       }),
     );